requirements.txt
crewai
crewai-tools
langgraph
langchain
langchain-community
duckduckgo-search
langchain-openai
pydantic

sphinx
sphinx-rtd-theme
python-dotenv
sphinx-autobuild
rich

streamlit
streamlit-agraph
litellm

arxiv
requests
pytest
